Create README
# OpenGl-
Mini Project in Computer Graphics using OpenGl.
Computer Graphics is concerned with all aspects of producing pictures or images using a computer. Here we use a particular graphics system, OpenGL which has become a widely accepted standard for developing graphics applications. A 2D OpenGL animation that demonstrates the use of glPushMatrix and glPopMatrix to implement hierarchical modelling entitled as  A Scenic view of Windmill. It is coded in a such way that when the Windmill is rotating, we can vary till 3 speed levels and  also to give some realistic feel we have also added a couple of cars which are in moving condition and you can also witness the same in both day and night respectively. I'm sure that you already Visualizing the project and further let's go deep into the project and gain some more knowledge about this.